Due to Rectangle2d extension in Sprite need to change primitives in methods.

+++ src/main/java/com/tarasek/mateusz/Bomb.java
@@ -21,7 +21,7 @@ public class Bomb extends Sprite implements SharedVariables{
     }
 
     void moveDownwards(){
-        setY(getY() + BOMB_SPEED);
+        setY((int)getY() + BOMB_SPEED);
     }
 
     void setDestroyed(boolean destroyed){
